+++ database/migrations/20211201192240_add_node_to_bundles_table.php
@@ -1,0 +1,21 @@
+<?php
+
+use Movim\Migration;
+use Illuminate\Database\Schema\Blueprint;
+
+class AddNodeToBundlesTable extends Migration
+{
+    public function up()
+    {
+        $this->schema->table('bundles', function (Blueprint $table) {
+            $table->string('node')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        $this->schema->table('bundles', function (Blueprint $table) {
+            $table->dropColumn('node');
+        });
+    }
+}